Completed Code

+++ dfp.js
@@ -1,15 +1,13 @@
 const fs = require('fs');
 
 function parseFile(indata, outdata, delimiter = ';') {
-  if (!fs.existsSync(indata)) {
-    return -1;
-  }
+  
+  // Return -1 if input file does not exist
+  if (!fs.existsSync(indata)) return -1;
 
   // Delete the output file if it exists
-  if (fs.existsSync(outdata)) {
-    fs.unlinkSync(outdata);
-  }
-
+  if (fs.existsSync(outdata)) fs.unlinkSync(outdata);
+  
   let recordCount = 0;
   try {
     // Read input file
@@ -19,7 +17,6 @@ function parseFile(indata, outdata, delimiter = ';') {
     const lines = fileContent.split(/\n/);
     
     // Run through each line and trim
-
     for (let i = 1; i < lines.length; i++) {
       const line = lines[i].trim();
       if (line === '') continue;
@@ -28,12 +25,12 @@ function parseFile(indata, outdata, delimiter = ';') {
 
       // Shorten review to 20 characters
       const shortReview = review.substring(0, 20);
-      console.log(shortReview)
 
       // Add new line to output file using append (Corrected line)
       fs.appendFileSync(outdata, `${sentiment}${delimiter}${shortReview}\n`, 'utf-8'); 
       recordCount++;
     }
+    // Returns -1 if there is an error in the try block
   } catch (err) {
     console.error('Error parsing file:', err); 
     return -1; 
@@ -42,9 +39,8 @@ function parseFile(indata, outdata, delimiter = ';') {
   // Return total number of records
   return recordCount;
 }
-  
-parseFile('./datafile.csv', './outputfile.csv')
 
+parseFile('./datafile.csv' , './outputfile.csv')
 
 // Leave this code here for the automated tests
 module.exports = {
